Move adapters into corresponding packages
